Expose current service/method in the Context
Add EndpointInfo() method to pull the ServiceInfo and MethodInfo of the
current endpoint out of the Context.

Defer generation of the context to after the service/method have been
identified. This prevents unnecessary setup operations (including
AppEngine binding), until after the request is verified to be
referencing a valid endpoint.